Add a TravisCi file to build on a push to master and release JARs
To make these easier to play around with, it would be nice if the
JARs were provided.

This commit adds a `.travis.yml` file that will build the project
and upload all JARs to the github releases page.

You'll need to create a github personal access token with `public_repo`
access and submit this as a secret environment variable on the travis job.

**note:** I commented out anything relating to keycloak-testsuite as I was
unable to find a maven repository serving those dependencies